requirements.txt.txt
python-telegram-bot
openpyxl
pandas